requirements.txt
absl-py==1.2.0
altair==4.2.0
antlr4-python3-runtime==4.9.3
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==22.3.0
blinker==1.5
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.1.0
commonmark==0.9.1
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cycler==0.11.0
decorator==5.1.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@717ab9f0aeca216a2f800e43d705766251ba3a55
entrypoints==0.4
fairscale==0.4.6
filelock==3.8.0
fonttools==4.37.1
future==0.18.2
fvcore==0.1.5.post20220512
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
hydra-core==1.2.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.12.0
importlib-resources==5.9.0
iopath==0.1.9
Jinja2==3.1.2
jsonschema==4.15.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mypy-extensions==0.4.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
pathspec==0.10.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
portalocker==2.5.1
protobuf==3.19.4
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.0b1
pydot==1.4.2
Pygments==2.13.0
Pympler==1.0.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
semver==2.13.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
soupsieve==2.3.2.post1
streamlit==1.12.2
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
timm==0.6.7
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
typed-ast==1.5.4
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
tzdata==2022.2
tzlocal==4.2
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
yacs==0.1.8
zipp==3.8.1
